BUG: join does not need a wildcard and specifying one breaks on windows
On windows, if you try to run the server, you get the error

 unable to locate 'C:\src\VolView\src\io\itk-dicom\web-build\dicom*' glob

 ERROR in unable to locate
 'C:\src\VolView\src\io\itk-dicom\web-build\dicom*' glob

This is used to copy files for the dist - copying files that start
with dicom in the web-build dir.  However, it fails on windows, and every
file in that dir starts with dicom, so you can just specify the top
level dir, web-build.  Confirmed to work on Windows.