fix(Oblique3D): change to use a composite representation proxy
Create a new representation proxy called vtkMultiSliceRepresentationProxy,
which will represent the outline box of the source imagedata and the
three vtkResliceRepresentationProxy reslice proxies internally.

Update VtkObliqueThreeView to use this new proxy.

This is required since the architecture of VolView is designed such that
in any given "view", there can be exactly one representation proxy that
can be used for each source type (e.g. vtkImageData).